+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+- [Core] Remove legacy `MGLMapboxAccessToken`. Support `MBXAccessToken` key in Info.plist exclusively.
 - [Core] Updated to Xcode 14.1 minimum version
 - [Core] Updated deployment target to iOS 12
 - [Address Autofill] Suggestions no longer perform a `retrieve` call.